Ajout de la condition de producteur actif à chaque fois pour permettre le stop de l'execution

+++ Projet_ProdCons/src/prodcons/v1/Consommateur.java
@@ -4,10 +4,12 @@ public class Consommateur extends Thread{
 	ProdConsBuffer pcbuffer;
 	Message msg;
 	int delay;
+	int nProd;
 
 	public Consommateur(ProdConsBuffer pcbuffer) {
 		this.pcbuffer = pcbuffer;
 		this.delay = delay;
+		this.nProd = pcbuffer.nProd;
 		start();
 	}
 	
@@ -15,7 +17,8 @@ public class Consommateur extends Thread{
 		try {
 			while(true) {
 				msg = pcbuffer.get(); //makes a consummer not die during it's execution. We'll have to figure out the stopping condition
-				if(pcbuffer.nmsg() == 0) //When there's no message in the buffer
+				nProd--;
+				if(pcbuffer.nmsg() == 0 && nProd == 0) //When there's no message in the buffer
 					System.exit(0);  // kill all hahaha
 			}	
 		} catch (InterruptedException e) {
